+++ expo-ai-chatbot/babel.config.js
@@ -10,6 +10,7 @@ module.exports = (api) => {
     ],
     plugins: [
       "@babel/plugin-proposal-export-namespace-from",
+      "@babel/plugin-transform-class-static-block",
       "react-native-reanimated/plugin",
     ],
   };